Remove Cargo.lock from .gitignore
As the comment in the .gitignore says, it should only be ignored for
libraries. As we're creating an executable rather than a library, we
shouldn't ignore the lock file. It also seems to cause issues when
building a Flatpak on Ubuntu machines, so let's just add the file to our
repo